Implement diagonal_loop
