+++ ket/include/ket/mpi/gate/controlled_phase_shift.hpp
@@ -1,724 +1,5 @@
 #ifndef KET_USE_DIAGONAL_LOOP
-# include <ket/mpi/gate/controlled_phase_shift.hpp.old>
+# include <ket/mpi/gate/detail/controlled_phase_shift_standard.hpp>
 #else // KET_USE_DIAGONAL_LOOP
-//
-#ifndef KET_MPI_GATE_CONTROLLED_PHASE_SHIFT_HPP
-# define KET_MPI_GATE_CONTROLLED_PHASE_SHIFT_HPP
-
-# include <complex>
-# include <vector>
-# include <ios>
-# include <sstream>
-
-# include <boost/range/value_type.hpp>
-
-# include <yampi/environment.hpp>
-# include <yampi/datatype_base.hpp>
-# include <yampi/communicator.hpp>
-
-# include <ket/qubit.hpp>
-# include <ket/control.hpp>
-# include <ket/qubit_io.hpp>
-# include <ket/control_io.hpp>
-# include <ket/gate/controlled_phase_shift.hpp>
-# include <ket/utility/begin.hpp>
-# include <ket/mpi/qubit_permutation.hpp>
-# include <ket/mpi/utility/general_mpi.hpp>
-# include <ket/mpi/utility/logger.hpp>
-# include <ket/mpi/gate/page/controlled_phase_shift.hpp>
-# include <ket/mpi/page/is_on_page.hpp>
-
-
-// TODO: implement vector-support (KET_PREFER_POINTER_TO_VECTOR_ITERATOR)
-namespace ket
-{
-  namespace mpi
-  {
-    namespace gate
-    {
-      // controlled_phase_shift_coeff
-      namespace controlled_phase_shift_detail
-      {
-        template <
-          typename MpiPolicy, typename ParallelPolicy,
-          typename RandomAccessRange, typename Complex,
-          typename StateInteger, typename BitInteger, typename Allocator>
-        inline RandomAccessRange& controlled_phase_shift_coeff(
-          MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-          RandomAccessRange& local_state,
-          Complex const& phase_coefficient,
-          ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-          ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-          ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-          yampi::communicator const& communicator,
-          yampi::environment const& environment)
-        {
-          if (::ket::mpi::page::is_on_page(target_qubit, local_state, permutation))
-          {
-            if (::ket::mpi::page::is_on_page(control_qubit.qubit(), local_state, permutation))
-              return ::ket::mpi::gate::page::controlled_phase_shift_coeff_tcp(
-                mpi_policy, parallel_policy,
-                local_state, phase_coefficient, target_qubit, control_qubit,
-                permutation);
-
-            return ::ket::mpi::gate::page::controlled_phase_shift_coeff_tp(
-              mpi_policy, parallel_policy,
-              local_state, phase_coefficient, target_qubit, control_qubit,
-              permutation, communicator.rank(environment));
-          }
-          else if (::ket::mpi::page::is_on_page(control_qubit.qubit(), local_state, permutation))
-            return ::ket::mpi::gate::page::controlled_phase_shift_coeff_cp(
-              mpi_policy, parallel_policy,
-              local_state, phase_coefficient, target_qubit, control_qubit,
-              permutation, communicator.rank(environment));
-
-          auto local_state_first = ::ket::utility::begin(local_state);
-
-          ::ket::mpi::utility::diagonal_loop(
-            mpi_policy, parallel_policy,
-            local_state, permutation, communicator, environment, target_qubit,
-            [](StateInteger const) { },
-            [local_state_first, &phase_coefficient](StateInteger const index)
-            { *(local_state_first+index) *= phase_coefficient; },
-            control_qubit);
-
-          return local_state;
-        }
-      } // namespace controlled_phase_shift_detail
-
-      template <
-        typename MpiPolicy, typename ParallelPolicy,
-        typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& controlled_phase_shift_coeff(
-        MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>&,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        auto output_string_stream = std::ostringstream{"Cphase(coeff) ", std::ios_base::ate};
-        output_string_stream << phase_coefficient << ' ' << target_qubit << ' ' << control_qubit;
-        ::ket::mpi::utility::log_with_time_guard<char> print{output_string_stream.str(), environment};
-
-        return ::ket::mpi::gate::controlled_phase_shift_detail::controlled_phase_shift_coeff(
-          mpi_policy, parallel_policy,
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          communicator, environment);
-      }
-
-      template <
-        typename MpiPolicy, typename ParallelPolicy,
-        typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& controlled_phase_shift_coeff(
-        MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>&,
-        yampi::datatype_base<DerivedDatatype> const&,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        auto output_string_stream = std::ostringstream{"Cphase(coeff) ", std::ios_base::ate};
-        output_string_stream << phase_coefficient << ' ' << target_qubit << ' ' << control_qubit;
-        ::ket::mpi::utility::log_with_time_guard<char> print{output_string_stream.str(), environment};
-
-        return ::ket::mpi::gate::controlled_phase_shift_detail::controlled_phase_shift_coeff(
-          mpi_policy, parallel_policy,
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          communicator, environment);
-      }
-
-      template <
-        typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& controlled_phase_shift_coeff(
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::controlled_phase_shift_coeff(
-          ::ket::mpi::utility::policy::make_general_mpi(),
-          ::ket::utility::policy::make_sequential(),
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          buffer, communicator, environment);
-      }
-
-      template <
-        typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& controlled_phase_shift_coeff(
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::datatype_base<DerivedDatatype> const& datatype,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::controlled_phase_shift_coeff(
-          ::ket::mpi::utility::policy::make_general_mpi(),
-          ::ket::utility::policy::make_sequential(),
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          buffer, datatype, communicator, environment);
-      }
-
-      template <
-        typename ParallelPolicy, typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& controlled_phase_shift_coeff(
-        ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::controlled_phase_shift_coeff(
-          ::ket::mpi::utility::policy::make_general_mpi(), parallel_policy,
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          buffer, communicator, environment);
-      }
-
-      template <
-        typename ParallelPolicy, typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& controlled_phase_shift_coeff(
-        ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::datatype_base<DerivedDatatype> const& datatype,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::controlled_phase_shift_coeff(
-          ::ket::mpi::utility::policy::make_general_mpi(), parallel_policy,
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          buffer, datatype, communicator, environment);
-      }
-
-      namespace controlled_phase_shift_detail
-      {
-        template <
-          typename MpiPolicy, typename ParallelPolicy,
-          typename RandomAccessRange, typename Complex,
-          typename StateInteger, typename BitInteger, typename Allocator>
-        inline RandomAccessRange& adj_controlled_phase_shift_coeff(
-          MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-          RandomAccessRange& local_state,
-          Complex const& phase_coefficient,
-          ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-          ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-          ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-          yampi::communicator const& communicator,
-          yampi::environment const& environment)
-        {
-          using std::conj;
-          return ::ket::mpi::gate::controlled_phase_shift_detail::controlled_phase_shift_coeff(
-            mpi_policy, parallel_policy,
-            local_state, conj(phase_coefficient), target_qubit, control_qubit, permutation,
-            communicator, environment);
-        }
-      } // namespace controlled_phase_shift_detail
-
-      template <
-        typename MpiPolicy, typename ParallelPolicy,
-        typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& adj_controlled_phase_shift_coeff(
-        MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>&,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        auto output_string_stream = std::ostringstream{"Adj(Cphase(coeff)) ", std::ios_base::ate};
-        output_string_stream << phase_coefficient << ' ' << target_qubit << ' ' << control_qubit;
-        ::ket::mpi::utility::log_with_time_guard<char> print{output_string_stream.str(), environment};
-
-        return ::ket::mpi::gate::controlled_phase_shift_detail::adj_controlled_phase_shift_coeff(
-          mpi_policy, parallel_policy,
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          communicator, environment);
-      }
-
-      template <
-        typename MpiPolicy, typename ParallelPolicy,
-        typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& adj_controlled_phase_shift_coeff(
-        MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>&,
-        yampi::datatype_base<DerivedDatatype> const&,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        auto output_string_stream = std::ostringstream{"Adj(Cphase(coeff)) ", std::ios_base::ate};
-        output_string_stream << phase_coefficient << ' ' << target_qubit << ' ' << control_qubit;
-        ::ket::mpi::utility::log_with_time_guard<char> print{output_string_stream.str(), environment};
-
-        return ::ket::mpi::gate::controlled_phase_shift_detail::adj_controlled_phase_shift_coeff(
-          mpi_policy, parallel_policy,
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          communicator, environment);
-      }
-
-      template <
-        typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& adj_controlled_phase_shift_coeff(
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::adj_controlled_phase_shift_coeff(
-          ::ket::mpi::utility::policy::make_general_mpi(),
-          ::ket::utility::policy::make_sequential(),
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          buffer, communicator, environment);
-      }
-
-      template <
-        typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& adj_controlled_phase_shift_coeff(
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::datatype_base<DerivedDatatype> const& datatype,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::adj_controlled_phase_shift_coeff(
-          ::ket::mpi::utility::policy::make_general_mpi(),
-          ::ket::utility::policy::make_sequential(),
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          buffer, datatype, communicator, environment);
-      }
-
-      template <
-        typename ParallelPolicy, typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& adj_controlled_phase_shift_coeff(
-        ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::adj_controlled_phase_shift_coeff(
-          ::ket::mpi::utility::policy::make_general_mpi(), parallel_policy,
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          buffer, communicator, environment);
-      }
-
-      template <
-        typename ParallelPolicy, typename RandomAccessRange, typename Complex,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& adj_controlled_phase_shift_coeff(
-        ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Complex const& phase_coefficient,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::datatype_base<DerivedDatatype> const& datatype,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::adj_controlled_phase_shift_coeff(
-          ::ket::mpi::utility::policy::make_general_mpi(), parallel_policy,
-          local_state, phase_coefficient, target_qubit, control_qubit, permutation,
-          buffer, datatype, communicator, environment);
-      }
-
-      // phase_shift
-      namespace controlled_phase_shift_detail
-      {
-        template <
-          typename MpiPolicy, typename ParallelPolicy,
-          typename RandomAccessRange, typename Real,
-          typename StateInteger, typename BitInteger, typename Allocator>
-        inline RandomAccessRange& controlled_phase_shift(
-          MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-          RandomAccessRange& local_state,
-          Real const phase,
-          ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-          ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-          ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-          yampi::communicator const& communicator,
-          yampi::environment const& environment)
-        {
-          using complex_type = typename boost::range_value<RandomAccessRange>::type;
-          return ::ket::mpi::gate::controlled_phase_shift_detail::controlled_phase_shift_coeff(
-            mpi_policy, parallel_policy,
-            local_state, ::ket::utility::exp_i<complex_type>(phase),
-            target_qubit, control_qubit, permutation,
-            communicator, environment);
-        }
-      } // namespace controlled_phase_shift_detail
-
-      template <
-        typename MpiPolicy, typename ParallelPolicy,
-        typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& controlled_phase_shift(
-        MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>&,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        auto output_string_stream = std::ostringstream{"Cphase ", std::ios_base::ate};
-        output_string_stream << phase << ' ' << target_qubit << ' ' << control_qubit;
-        ::ket::mpi::utility::log_with_time_guard<char> print{output_string_stream.str(), environment};
-
-        return ::ket::mpi::gate::controlled_phase_shift_detail::controlled_phase_shift(
-          mpi_policy, parallel_policy,
-          local_state, phase, target_qubit, control_qubit, permutation,
-          communicator, environment);
-      }
-
-      template <
-        typename MpiPolicy, typename ParallelPolicy,
-        typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& controlled_phase_shift(
-        MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>&,
-        yampi::datatype_base<DerivedDatatype> const&,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        auto output_string_stream = std::ostringstream{"Cphase ", std::ios_base::ate};
-        output_string_stream << phase << ' ' << target_qubit << ' ' << control_qubit;
-        ::ket::mpi::utility::log_with_time_guard<char> print{output_string_stream.str(), environment};
-
-        return ::ket::mpi::gate::controlled_phase_shift_detail::controlled_phase_shift(
-          mpi_policy, parallel_policy,
-          local_state, phase, target_qubit, control_qubit, permutation,
-          communicator, environment);
-      }
-
-      template <
-        typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& controlled_phase_shift(
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::controlled_phase_shift(
-          ::ket::mpi::utility::policy::make_general_mpi(),
-          ::ket::utility::policy::make_sequential(),
-          local_state, phase, target_qubit, control_qubit, permutation,
-          buffer, communicator, environment);
-      }
-
-      template <
-        typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& controlled_phase_shift(
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::datatype_base<DerivedDatatype> const& datatype,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::controlled_phase_shift(
-          ::ket::mpi::utility::policy::make_general_mpi(),
-          ::ket::utility::policy::make_sequential(),
-          local_state, phase, target_qubit, control_qubit, permutation,
-          buffer, datatype, communicator, environment);
-      }
-
-      template <
-        typename ParallelPolicy, typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& controlled_phase_shift(
-        ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::controlled_phase_shift(
-          ::ket::mpi::utility::policy::make_general_mpi(), parallel_policy,
-          local_state, phase, target_qubit, control_qubit, permutation,
-          buffer, communicator, environment);
-      }
-
-      template <
-        typename ParallelPolicy, typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& controlled_phase_shift(
-        ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::datatype_base<DerivedDatatype> const& datatype,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::controlled_phase_shift(
-          ::ket::mpi::utility::policy::make_general_mpi(), parallel_policy,
-          local_state, phase, target_qubit, control_qubit, permutation,
-          buffer, datatype, communicator, environment);
-      }
-
-      namespace controlled_phase_shift_detail
-      {
-        template <
-          typename MpiPolicy, typename ParallelPolicy,
-          typename RandomAccessRange, typename Real,
-          typename StateInteger, typename BitInteger, typename Allocator>
-        inline RandomAccessRange& adj_controlled_phase_shift(
-          MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-          RandomAccessRange& local_state,
-          Real const phase,
-          ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-          ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-          ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-          yampi::communicator const& communicator,
-          yampi::environment const& environment)
-        {
-          return ::ket::mpi::gate::controlled_phase_shift_detail::controlled_phase_shift(
-            mpi_policy, parallel_policy,
-            local_state, -phase, target_qubit, control_qubit, permutation,
-            communicator, environment);
-        }
-      } // namespace controlled_phase_shift_detail
-
-      template <
-        typename MpiPolicy, typename ParallelPolicy,
-        typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& adj_controlled_phase_shift(
-        MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>&,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        auto output_string_stream = std::ostringstream{"Adj(Cphase) ", std::ios_base::ate};
-        output_string_stream << phase << ' ' << target_qubit << ' ' << control_qubit;
-        ::ket::mpi::utility::log_with_time_guard<char> print{output_string_stream.str(), environment};
-
-        return ::ket::mpi::gate::controlled_phase_shift_detail::adj_controlled_phase_shift(
-          mpi_policy, parallel_policy,
-          local_state, phase, target_qubit, control_qubit, permutation,
-          communicator, environment);
-      }
-
-      template <
-        typename MpiPolicy, typename ParallelPolicy,
-        typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& adj_controlled_phase_shift(
-        MpiPolicy const mpi_policy, ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>&,
-        yampi::datatype_base<DerivedDatatype> const&,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        auto output_string_stream = std::ostringstream{"Adj(Cphase) ", std::ios_base::ate};
-        output_string_stream << phase << ' ' << target_qubit << ' ' << control_qubit;
-        ::ket::mpi::utility::log_with_time_guard<char> print{output_string_stream.str(), environment};
-
-        return ::ket::mpi::gate::controlled_phase_shift_detail::adj_controlled_phase_shift(
-          mpi_policy, parallel_policy,
-          local_state, phase, target_qubit, control_qubit, permutation,
-          communicator, environment);
-      }
-
-      template <
-        typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& adj_controlled_phase_shift(
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::adj_controlled_phase_shift(
-          ::ket::mpi::utility::policy::make_general_mpi(),
-          ::ket::utility::policy::make_sequential(),
-          local_state, phase, target_qubit, control_qubit, permutation,
-          buffer, communicator, environment);
-      }
-
-      template <
-        typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& adj_controlled_phase_shift(
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::datatype_base<DerivedDatatype> const& datatype,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::adj_controlled_phase_shift(
-          ::ket::mpi::utility::policy::make_general_mpi(),
-          ::ket::utility::policy::make_sequential(),
-          local_state, phase, target_qubit, control_qubit, permutation,
-          buffer, datatype, communicator, environment);
-      }
-
-      template <
-        typename ParallelPolicy, typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator>
-      inline RandomAccessRange& adj_controlled_phase_shift(
-        ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::adj_controlled_phase_shift(
-          ::ket::mpi::utility::policy::make_general_mpi(), parallel_policy,
-          local_state, phase, target_qubit, control_qubit, permutation,
-          buffer, communicator, environment);
-      }
-
-      template <
-        typename ParallelPolicy, typename RandomAccessRange, typename Real,
-        typename StateInteger, typename BitInteger,
-        typename Allocator, typename BufferAllocator, typename DerivedDatatype>
-      inline RandomAccessRange& adj_controlled_phase_shift(
-        ParallelPolicy const parallel_policy,
-        RandomAccessRange& local_state,
-        Real const phase,
-        ::ket::qubit<StateInteger, BitInteger> const target_qubit,
-        ::ket::control< ::ket::qubit<StateInteger, BitInteger> > const control_qubit,
-        ::ket::mpi::qubit_permutation<StateInteger, BitInteger, Allocator>& permutation,
-        std::vector<typename boost::range_value<RandomAccessRange>::type, BufferAllocator>& buffer,
-        yampi::datatype_base<DerivedDatatype> const& datatype,
-        yampi::communicator const& communicator,
-        yampi::environment const& environment)
-      {
-        return ::ket::mpi::gate::adj_controlled_phase_shift(
-          ::ket::mpi::utility::policy::make_general_mpi(), parallel_policy,
-          local_state, phase, target_qubit, control_qubit, permutation,
-          buffer, datatype, communicator, environment);
-      }
-    } // namespace gate
-  } // namespace mpi
-} // namespace ket
-
-
-#endif // KET_MPI_GATE_CONTROLLED_PHASE_SHIFT_HPP
-//
+# include <ket/mpi/gate/detail/controlled_phase_shift_diagonal.hpp>
 #endif // KET_USE_DIAGONAL_LOOP
