+++ ket/include/ket/mpi/gate/detail/controlled_phase_shift_standard.hpp
@@ -1,5 +1,5 @@
-#ifndef KET_MPI_GATE_CONTROLLED_PHASE_SHIFT_HPP
-# define KET_MPI_GATE_CONTROLLED_PHASE_SHIFT_HPP
+#ifndef KET_MPI_GATE_DETAIL_CONTROLLED_PHASE_SHIFT_STANDARD_HPP
+# define KET_MPI_GATE_DETAIL_CONTROLLED_PHASE_SHIFT_STANDARD_HPP
 
 # include <boost/config.hpp>
 
@@ -895,4 +895,4 @@ namespace ket
 } // namespace ket
 
 
-#endif // KET_MPI_GATE_CONTROLLED_PHASE_SHIFT_HPP
+#endif // KET_MPI_GATE_DETAIL_CONTROLLED_PHASE_SHIFT_STANDARD_HPP
