+++ ket/include/ket/mpi/gate/projective_measurement.hpp
@@ -33,8 +33,6 @@
 # include <ket/mpi/page/is_on_page.hpp>
 
 
-// TODO: implement vector-support (KET_PREFER_POINTER_TO_VECTOR_ITERATOR)
-// avoid direct calling ket::gate::projective_measurement_detail::*
 namespace ket
 {
   namespace mpi
